some infrastructure for using Csound's "virtual keyboard"
MIDI signal dispatcher (intended for a future organ-stop thing) and a
simple, not intended to be very realistic, piano patch based on gbuzz.
Velocity will come from controller 7 if that is nonzero, since the
virtual keyboard sends maximum velocity all the time.  Note keyboard map
file; it is undocumented and annoying that program numbers used by
the keyboard MUST be consecutive starting from 1.